requirements.txt
# Python dependencies
streamlit
psycopg2-binary
pgvector
sqlparse
pandas
google-generativeai 
sqlalchemy
python-dotenv
